Move PRODUCT_FEATURE_RX, TOOL_TIMEOUT, and ENCODING to settings
These changes moved the values from tokenstat to settings, and
moved code to cmd_utils from slurmcltd_prolog.++++++++++
